% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermole-package.R
\docType{package}
\name{clustermole-package}
\alias{clustermole}
\alias{clustermole-package}
\title{clustermole: Unbiased Single-Cell Transcriptomic Data Cell Type Identification}
\description{
Assignment of cell type labels to single-cell RNA sequencing (scRNA-seq) clusters is often a time-consuming process that involves manual inspection of the cluster marker genes complemented with a detailed literature search. This is especially challenging when unexpected or poorly described populations are present. The clustermole R package provides methods to query thousands of human and mouse cell identity markers sourced from a variety of databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://igordot.github.io/clustermole/}
  \item Report bugs at \url{https://github.com/igordot/clustermole/issues}
}

}
\author{
\strong{Maintainer}: Igor Dolgalev \email{igor.dolgalev@nyumc.org} (\href{https://orcid.org/0000-0003-4451-126X}{ORCID})

Authors:
\itemize{
  \item Igor Dolgalev \email{igor.dolgalev@nyumc.org} (\href{https://orcid.org/0000-0003-4451-126X}{ORCID})
}

}
\keyword{internal}
