% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_gmt}
\alias{read_gmt}
\title{Read a GMT file into a data frame}
\usage{
read_gmt(file, geneset_label = "celltype", gene_label = "gene")
}
\arguments{
\item{file}{A file path, URL, or connection.}

\item{geneset_label}{Output column name for gene sets (GMT column 1).}

\item{gene_label}{Output column name for genes (GMT columns 3 onward).}
}
\value{
A data frame with gene sets and genes, one gene per row.
}
\description{
Read a GMT file into a data frame
}
\examples{
\dontrun{
gmt <- "http://software.broadinstitute.org/gsea/msigdb/supplemental/scsig.all.v1.0.symbols.gmt"
gmt_tbl <- read_gmt(gmt)
head(gmt_tbl)
}
}
