% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermole-package.R
\docType{package}
\name{clustermole-package}
\alias{clustermole}
\alias{clustermole-package}
\title{clustermole: Cell Type Marker Database for Single-Cell RNA-Seq Data}
\description{
Provides a meta-database of thousands of human and mouse cell identity markers curated from multiple sources, along with methods for cell type prediction based on marker gene overlaps or gene set enrichment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://igordot.github.io/clustermole/}
  \item Report bugs at \url{https://github.com/igordot/clustermole/issues}
}

}
\author{
\strong{Maintainer}: Igor Dolgalev \email{igor.dolgalev@nyumc.org} (\href{https://orcid.org/0000-0003-4451-126X}{ORCID})

Authors:
\itemize{
  \item Igor Dolgalev \email{igor.dolgalev@nyumc.org} (\href{https://orcid.org/0000-0003-4451-126X}{ORCID})
}

}
\keyword{internal}
