% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clustermole_enrichment}
\alias{clustermole_enrichment}
\title{Perform cell type enrichment for a given gene expression matrix}
\usage{
clustermole_enrichment(expr_mat, species)
}
\arguments{
\item{expr_mat}{expression matrix (logCPMs, logFPKMs, or logTPMs)}

\item{species}{species ("hs" for human or "mm" for mouse)}
}
\value{
a data frame of enrichment results
}
\description{
Perform cell type enrichment for a given gene expression matrix
}
