% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clustermole_enrichment}
\alias{clustermole_enrichment}
\title{Retrieve the cell type markers data frame}
\usage{
clustermole_enrichment(expr_mat, species)
}
\arguments{
\item{expr_mat}{normalized expression matrix (log-scale)}

\item{species}{species ("hs" or "mm")}
}
\value{
a data frame of enrichment results
}
\description{
Retrieve the cell type markers data frame
}
