% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clustermole_enrichment}
\alias{clustermole_enrichment}
\title{Perform cell type enrichment for a given gene expression matrix}
\usage{
clustermole_enrichment(expr_mat, species)
}
\arguments{
\item{expr_mat}{Expression matrix (logCPMs, logFPKMs, or logTPMs) with genes as rows.}

\item{species}{Species: "hs" for human or "mm" for mouse.}
}
\value{
A data frame of enrichment results.
}
\description{
Perform cell type enrichment for a given gene expression matrix
}
\examples{
# my_enrichment <- clustermole_enrichment(expr_mat = my_expr_mat, species = "hs")
}
