% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clustermole_markers}
\alias{clustermole_markers}
\title{Retrieve a table of all cell type markers}
\usage{
clustermole_markers()
}
\value{
a data frame of markers with one gene per row
}
\description{
Retrieve a table of all cell type markers
}
\examples{
markers_tbl <- clustermole_markers()
}
