% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clustermole_markers}
\alias{clustermole_markers}
\title{Retrieve the available cell type markers}
\usage{
clustermole_markers(species = "hs")
}
\arguments{
\item{species}{Species for the appropriate gene symbol format: "hs" for human or "mm" for mouse.}
}
\value{
A data frame of markers with one gene per row.
}
\description{
Retrieve the available cell type markers
}
\examples{
markers <- clustermole_markers()
head(markers)
}
