% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clustermole_markers}
\alias{clustermole_markers}
\title{Retrieve the cell type markers data frame}
\usage{
clustermole_markers()
}
\value{
a data frame of markers with one gene per row
}
\description{
Retrieve the cell type markers data frame
}
