% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{clustermole_markers}
\alias{clustermole_markers}
\title{Available cell type markers}
\usage{
clustermole_markers(species = c("hs", "mm"))
}
\arguments{
\item{species}{Species: \code{hs} for human or \code{mm} for mouse.}
}
\value{
A data frame of cell type markers (one gene per row).
}
\description{
Retrieve the full list of cell type markers in the \code{clustermole} database.
}
\examples{
markers <- clustermole_markers()
head(markers)
}
